@@ -1,13 +1,12 @@
 plugins {
-  id 'org.springframework.boot' version '2.1.5.RELEASE'
+  id 'org.springframework.boot' version '2.5.4'
+  id 'io.spring.dependency-management' version '1.0.11.RELEASE'
   id 'java'
 }
 
-apply plugin: 'io.spring.dependency-management'
-
 group = 'com.jos.dem.security'
-version = '0.0.1-SNAPSHOT'
-sourceCompatibility = 11
+version = '1.0.0-SNAPSHOT'
+sourceCompatibility = 16
 
 configurations {
   compileOnly {
@@ -20,12 +19,18 @@ repositories {
 }
 
 dependencies {
-  implementation('org.springframework.boot:spring-boot-starter-webflux')
-  implementation('org.springframework.boot:spring-boot-starter-security')
-  implementation('org.springframework.boot:spring-boot-starter-thymeleaf')
-  implementation('org.springframework.boot:spring-boot-starter-data-mongodb-reactive')
-  compileOnly('org.projectlombok:lombok')
-  annotationProcessor('org.projectlombok:lombok')
-  testImplementation('org.springframework.boot:spring-boot-starter-test')
-  testImplementation('io.projectreactor:reactor-test')
+  implementation 'org.springframework.boot:spring-boot-starter-security'
+  implementation 'org.springframework.boot:spring-boot-starter-thymeleaf'
+  implementation 'org.springframework.boot:spring-boot-starter-webflux'
+  implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity5'
+  implementation 'org.springframework.boot:spring-boot-starter-data-mongodb-reactive'
+  compileOnly 'org.projectlombok:lombok'
+  annotationProcessor 'org.projectlombok:lombok'
+  testImplementation 'org.springframework.boot:spring-boot-starter-test'
+  testImplementation 'io.projectreactor:reactor-test'
+  testImplementation 'org.springframework.security:spring-security-test'
+}
+
+test {
+  useJUnitPlatform()
 }
