@@ -1,13 +1,10 @@
-package com.example.reactivewebfluxsecurity;
+package com.jos.dem.security;
 
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
-@RunWith(SpringRunner.class)
 @SpringBootTest
-public class DemoApplicationTests {
+class DemoApplicationTests {
 
 	@Test
 	public void contextLoads() {
